Refactor DenomTrace
